refactor(api): give stats their own services and per-endpoint suites
PlayerStatsAggregationService and TeamStatsAggregationService only held
the aggregate queries while the rest of the stats logic stayed in
PlayerService and TeamService. They are now PlayerStatsService and
TeamStatsService and own everything stats: the queries, the page
hydration, the api-model building and both endpoints. PlayerService and
TeamService are left with the VLR upsert helpers they started as.

Tests follow the same split. stats-aggregation and stats-freshness each
covered two endpoints at once; their assertions now live with the
endpoint they exercise — player leaderboard, ordering, pagination and
freshness in the player stats suite, the team equivalents in the team
suite. The player suite calls team endpoints only to set fixtures up, not
to assert on them.
